fix(src/nectarchain/acceptance_verification_package):
- deadtime.py: removed

- gui.py: now import deadtime.py from ../trr_test_suite

- ../trr_test_suite/deadtime.py: refactor to accomodate both 'trr' and 'av' tests, and GUI arguments

- source_type_labels.json: moved to ../trr_test_suite/resources
